refactor: 형식 맞추기
