+++ client/src/querys/useDeleteMutation.ts
@@ -1,0 +1,18 @@
+import { useMutation } from '@tanstack/react-query';
+
+// Todo useMutation Custom Hook으로 만들기
+
+// const useDeleteMutation = ({
+//   func,
+//   onSuccessFunc,
+// }: {
+//   func: <T>() => void;
+//   onSuccessFunc: <T>() => <U>;
+// }) => {
+//   const mutate = useMutation(func, {
+//     onSuccess: onSuccessFunc,
+//   });
+//   return mutate;
+// };
+
+// export default useDeleteMutation;
